添加宏定义 ltmax 和 ltmin，优化代码中最大值和最小值的计算；修复多个文件中获取单元格尺寸的逻辑，确保正确处理尺寸计算；更新项目文件以支持新的 Windows 目标平台版本。

+++ GridCtrl/CellTypes/GridCellCheck.cpp
@@ -35,7 +35,6 @@
 #include "pch.h"
 #include "../GridCell.h"
 #include "../GridCtrl.h"
-
 #include "GridCellCheck.h"
 
 
@@ -64,7 +63,7 @@ CSize CGridCellCheck::GetCellExtent(CDC* pDC)
     int nWidth = GetSystemMetrics(SM_CXHSCROLL) + 2*GetMargin();	
     CSize	cellSize = CGridCell::GetCellExtent(pDC);	
     cellSize.cx += nWidth;	
-    cellSize.cy = max (cellSize.cy, nWidth);	
+    cellSize.cy = max(cellSize.cy, (LONG)nWidth);
     return  cellSize;
 }
